Add comprehensive unit tests for calculator
